adding highlight container component, moving highlight components to subfolder
